+++ packages/design-vue/src/components/molecules/noise-tracking-controls/index.ts
@@ -1,0 +1,2 @@
+export * from './noise-tracking-controls.types';
+export { default } from './noise-tracking-controls.vue';