+++ packages/git-utils/src/index.ts
@@ -1,4 +1,6 @@
 export * from './git';
 export * from './git-analyzer';
 export * from './git-analyzer.types';
+export * from './git-service';
+export * from './git-service.types';
 export * from './git.types';