+++ packages/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @taskin/api
 
+## 1.0.8
+
+### Patch Changes
+
+- @opentask/taskin-core@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes