+++ packages/git-utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @opentask/taskin-git-utils
 
+## 2.1.3
+
+### Patch Changes
+
+- Remove unnecessary install scripts that caused pnpm build script warnings
+
+  Removed `install` scripts from all packages that only printed echo messages. These scripts were unnecessary since packages are already pre-built and included in the published bundle. This eliminates the "Ignored build scripts" warning when installing taskin in external projects.
+
+- Updated dependencies
+  - @opentask/taskin-types@1.1.1
+
 ## 2.1.2
 
 ### Patch Changes