+++ packages/task-manager/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @opentask/taskin-task-manager
 
+## 2.0.1
+
+### Patch Changes
+
+- Remove unnecessary install scripts that caused pnpm build script warnings
+
+  Removed `install` scripts from all packages that only printed echo messages. These scripts were unnecessary since packages are already pre-built and included in the published bundle. This eliminates the "Ignored build scripts" warning when installing taskin in external projects.
+
+- Updated dependencies
+  - @opentask/taskin-types@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes