+++ packages/taskin-alias/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @opentask/taskin
 
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - taskin@3.0.3
+
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - taskin@3.0.2
+
 ## 2.0.11
 
 ### Patch Changes